Prevent 100% cpu core usage in Connect mode
When we receive a command to stream a track in Connect mode, the spotty deamon could go into a tight loop, using all of a cpu core's cycles. Sleep the thread for a few milliseconds on each iteration, as at this point we're done receiving data, awaiting the next command. This brings CPU usage down to single digit percent. This is particularly important for lower power devices such as LMS running on Raspberry Pi.